requirements/base.txt
#
#   Django
#

Django==1.11.6
django-configurations==2.0


#
#   Environment and path management
#
python-dotenv==0.7.1
django-environ==0.4.4

#   dependencies
six==1.11.0


#
#   Postgres
#

psycopg2==2.7.3.2


#
#   Timezone calculation
#

pytz==2017.
